misc fixes

+++ lib/timestamp/unixTimestamp.go
@@ -4,30 +4,30 @@ import (
 	"fmt"
 )
 
-func UnixShortDate(unixTime int) string {
+func UnixShortDate(unixTime int64) string {
 	return fmt.Sprintf("<t:%v:d>", unixTime)
 }
 
-func UnixLongDate(unixTime int) string {
+func UnixLongDate(unixTime int64) string {
 	return fmt.Sprintf("<t:%v:D>", unixTime)
 }
 
-func UnixShortTime(unixTime int) string {
+func UnixShortTime(unixTime int64) string {
 	return fmt.Sprintf("<t:%v:t>", unixTime)
 }
 
-func UnixLongTime(unixTime int) string {
+func UnixLongTime(unixTime int64) string {
 	return fmt.Sprintf("<t:%v:T>", unixTime)
 }
 
-func UnixShortFull(unixTime int) string {
+func UnixShortFull(unixTime int64) string {
 	return fmt.Sprintf("<t:%v:f>", unixTime)
 }
 
-func UnixLongFull(unixTime int) string {
+func UnixLongFull(unixTime int64) string {
 	return fmt.Sprintf("<t:%v:F>", unixTime)
 }
 
-func UnixRelativeTime(unixTime int) string {
+func UnixRelativeTime(unixTime int64) string {
 	return fmt.Sprintf("<t:%v:R>", unixTime)
 }
